Added basic map rendering functionality
Created the `map.lua` file to render a simple map. This is currenlty hardcoded into the class
but is designed that it can be extended down the line once ready.

Currently only draws the 32x32 sprite at the exact size and doesn't scale them up. That is a
future commit's problem.